Construct ProductReviewsTab within its block
